+++ AWS/connectors/__init__.py
@@ -88,7 +88,7 @@ class AbstractAwsConnector(AsyncConnector, metaclass=ABCMeta):
                     # and current timestamp ( when it was processed )
                     processing_end = time.time()
                     for message_timestamp in messages_timestamp:
-                        EVENTS_LAG.labels(intake_key=self.configuration.intake_key).observe(
+                        EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(
                             processing_end - (message_timestamp / 1000)
                         )