+++ Duo/duo/connector.py
@@ -148,7 +148,7 @@ class DuoLogsConsumer(Thread):
                 events_lag = current_timestamp - most_recent_timestamp
                 EVENTS_LAG.labels(
                     intake_key=self.connector.configuration.intake_key, type=self._log_type.value
-                ).observe(events_lag)
+                ).set(events_lag)
 
             # Add `eventtype` field
             for event in events: