+++ Fastly/fastly_waf/connector_fastly_waf.py
@@ -116,7 +116,7 @@ class FastlyWAFConnector(Connector):
 
         now = datetime.datetime.now(datetime.timezone.utc)
         current_lag = now - most_recent_date_seen
-        EVENTS_LAG.labels(intake_key=self.configuration.intake_key).observe(int(current_lag.total_seconds()))
+        EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(int(current_lag.total_seconds()))
 
     def next_batch(self):
         # save the starting time