+++ Github/github_modules/audit_log_trigger.py
@@ -253,7 +253,7 @@ class AuditLogConnector(Connector):
                 # compute the lag
                 now = time.time()
                 current_lag = now - self.last_ts / 1000
-                EVENTS_LAG.labels(intake_key=self.configuration.intake_key).observe(int(current_lag))
+                EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(int(current_lag))
             else:
                 self.log(
                     message="No events to forward ",