+++ Netskope/netskope_modules/connector_pull_events_v2.py
@@ -102,7 +102,7 @@ class NetskopeEventConsumer(Thread):
         if most_recent_timestamp > 0:
             now = time.time()
             current_lag = now - most_recent_timestamp
-            EVENTS_LAG.labels(intake_key=self.connector.configuration.intake_key, type=self.name).observe(
+            EVENTS_LAG.labels(intake_key=self.connector.configuration.intake_key, type=self.name).set(
                 int(current_lag)
             )