+++ SkyhighSecurity/gateway_cloud_services/trigger_skyhigh_security_swg.py
@@ -69,7 +69,7 @@ class EventCollector(Thread):
             message=f"Current lag {int(current_lag.total_seconds())} seconds.",
             level="info",
         )
-        EVENTS_LAG.labels(intake_key=self.configuration.intake_key).observe(int(current_lag.total_seconds()))
+        EVENTS_LAG.labels(intake_key=self.configuration.intake_key).set(int(current_lag.total_seconds()))
 
         if self.end_date >= now:
             difference = self.end_date - now
@@ -107,7 +107,7 @@ class EventCollector(Thread):
             message=f"Skyhigh API took {time_elapsed} to answer our query",
             level="info",
         )
-        COLLECT_EVENTS_DURATION.labels(intake_key=self.configuration.intake_key).observe(
+        COLLECT_EVENTS_DURATION.labels(intake_key=self.configuration.intake_key).set(
             int(time_elapsed.total_seconds())
         )